Add more refactors to the TS SDK
